audio: Fix incall issues caused by playback
If we are in a call but get a playback sound, the devices are changed
from voice-speaker to speaker. This means we have different routing and
incall sound stops working.

If you want to reproduce this turn on 'Touch sounds' and press the
speaker button while in a call. This will trigger to play a sound.

Change-Id: I717c09a2bee091e31645d7e8a01bbe0ef8b919f1